+++ packages/insomnia-smoke-test/server/index.ts
@@ -10,14 +10,12 @@ import gitlabApi from './gitlab-api';
 import { root, schema } from './graphql';
 import { startGRPCServer } from './grpc';
 import { oauthRoutes } from './oauth';
-import { startPreReleaseGRPCServer } from './prerelease-grpc';
 import { startWebSocketServer } from './websocket';
 
 const app = express();
 const port = 4010;
 const httpsPort = 4011;
 const grpcPort = 50051;
-const grpcRouteGuidePort = 50052;
 
 app.get('/pets/:id', (req, res) => {
   res.status(200).send({ id: req.params.id });
@@ -61,27 +59,21 @@ app.use('/graphql', graphqlHTTP({
   graphiql: true,
 }));
 
-startGRPCServer(grpcPort).then(() => {
+startWebSocketServer(app.listen(port, () => {
+  console.log(`Listening at http://localhost:${port}`);
+  console.log(`Listening at ws://localhost:${port}`);
+}));
 
-  const server = app.listen(port, () => {
-    console.log(`Listening at http://localhost:${port}`);
-    console.log(`Listening at ws://localhost:${port}`);
-  });
+startWebSocketServer(createServer({
+  cert: readFileSync(join(__dirname, '../fixtures/certificates/localhost.pem')),
+  ca: readFileSync(join(__dirname, '../fixtures/certificates/rootCA.pem')),
+  key: readFileSync(join(__dirname, '../fixtures/certificates/localhost-key.pem')),
+  // Only allow connections using valid client certificates
+  requestCert: true,
+  rejectUnauthorized: true,
+}, app).listen(httpsPort, () => {
+  console.log(`Listening at https://localhost:${httpsPort}`);
+  console.log(`Listening at wss://localhost:${httpsPort}`);
+}));
 
-  const httpsServer = createServer({
-    cert: readFileSync(join(__dirname, '../fixtures/certificates/localhost.pem')),
-    ca: readFileSync(join(__dirname, '../fixtures/certificates/rootCA.pem')),
-    key: readFileSync(join(__dirname, '../fixtures/certificates/localhost-key.pem')),
-    // Only allow connections using valid client certificates
-    requestCert: true,
-    rejectUnauthorized: true,
-  }, app);
-  httpsServer.listen(httpsPort, () => {
-    console.log(`Listening at https://localhost:${httpsPort}`);
-    console.log(`Listening at wss://localhost:${httpsPort}`);
-  });
-
-  startPreReleaseGRPCServer(grpcRouteGuidePort);
-  startWebSocketServer(server);
-  startWebSocketServer(httpsServer);
-});
+startGRPCServer(grpcPort);
